# PERF: DataFrame.T