--- a/pandas/tests/frame/test_constructors.py
+++ b/pandas/tests/frame/test_constructors.py
@@ -70,13 +70,16 @@ MIXED_INT_DTYPES = [
 
 
 class TestDataFrameConstructors:
-    def test_constructor_from_2d_datetimearray(self):
+    def test_constructor_from_2d_datetimearray(self, using_array_manager):
         dti = date_range("2016-01-01", periods=6, tz="US/Pacific")
         dta = dti._data.reshape(3, 2)
 
         df = DataFrame(dta)
         expected = DataFrame({0: dta[:, 0], 1: dta[:, 1]})
         tm.assert_frame_equal(df, expected)
+        if not using_array_manager:
+            # GH#??? big performance hit if we de-consolidate
+            assert len(df._mgr.blocks) == 1
 
     def test_constructor_dict_with_tzaware_scalar(self):
         # GH#42505