--- a/pandas/tests/frame/test_constructors.py
+++ b/pandas/tests/frame/test_constructors.py
@@ -78,7 +78,7 @@ class TestDataFrameConstructors:
         expected = DataFrame({0: dta[:, 0], 1: dta[:, 1]})
         tm.assert_frame_equal(df, expected)
         if not using_array_manager:
-            # GH#??? big performance hit if we de-consolidate
+            # GH#44724 big performance hit if we de-consolidate
             assert len(df._mgr.blocks) == 1
 
     def test_constructor_dict_with_tzaware_scalar(self):