- [ ] closes #xxxx
- [x] tests added / passed
- [x] Ensure all linting tests pass, see [here](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit) for how to run them
- [ ] whatsnew entry

Perf regression caused by #44593 (not yet released, so no whatsnew)

```
from asv_bench.benchmarks.reshape import *
cls = ReshapeExtensionDtype
dtype = cls.params[0]
self = cls()
self.setup(dtype)

%timeit self.df.T
91.1 ms ± 1.75 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- master
163 µs ± 3.8 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- PR

```